refactor: view 영역 domain 의존도 분리
