+++ readme.md
@@ -41,8 +41,8 @@
 - [x] 엔트리 포인트 정의
 - [x] car static 제거
 - [x] 랜덤 값 contest에서 받아서 차량 움직이는 방식으로 리팩토링
-- [ ] 확률적 테스트 제거 후, 테스트 보완
-- [ ] view 영역 domain 의존도 분리
+- [x] 확률적 테스트 제거
+- [x] view 영역 domain 의존도 분리
 
 ## 고민한 부분
